- [Tests] Updated the fixture to create and resue the JotformClient instance throughout all tests.
- [Fix] Changed the GetUserUsage.Uploads property to long (from int), as I've got a failing test because the upload number went higher than int.

+++ Jotform/Endpoints/User/GetUserUsage.cs
@@ -35,7 +35,7 @@ public class GetUserUsageResponse
     /// Total disk space used for uploaded files. In bytes.
     /// </summary>
     [JsonPropertyName("uploads")]
-    public int Uploads { get; set; }
+    public long Uploads { get; set; }
 
     /// <summary>
     /// Number of mobile submissions received this month
